+++ src/storm/settings/SettingsManager.h
@@ -65,6 +65,11 @@ namespace storm {
             void setFromConfigurationFile(std::string const& configFilename);
             
             /*!
+             * Throws an exception with a nice error message indicating similar valid option names.
+             */
+            void handleUnknownOption(std::string const& optionName, bool isShort) const;
+            
+            /*!
              * This function prints a help message to the standard output. Optionally, a string can be given as a hint.
              * If it is 'all', the complete help is shown. Otherwise, the string is interpreted as a regular expression
              * and matched against the known modules and options to restrict the help output.