+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 1.3.x
 - Added support for multi-dimensional quantile queries
 - Allow to quickly check a benchmark from the [Quantitative Verification Benchmark Set](http://qcomp.org/benchmarks/) using the --qvbs option.
 - Added script resources/examples/download_qvbs.sh to download the QVBS.
+- If an option is unknown, Storm now prints a hint to similar option names.
 - Allow bounded types for constants in JANI.
 - Fixed sparse bisimulation of MDPs (which failed if all non-absorbing states in the quotient are initial)